Zero warnings!

+++ Client/Source/AudioCallback.h
@@ -35,8 +35,8 @@ public:
 	std::weak_ptr<MidiClocker> getClocker();
 
 	// Statistics
-	int numberOfUnderruns() const;
-	int currentBufferSize() const;
+	int64 numberOfUnderruns() const;
+	uint64 currentBufferSize() const;
 	int currentPacketSize() const;
 	uint64 currentPlayQueueSize() const;
 	int currentDiscardedPackageCounter() const;
